remove old class file
