+++ Source/MultichannelWidget/QismoManager.swift
@@ -1,15 +1,15 @@
 //
-//  QismoManager.swift
-//  BBChat
+//  QismoManager2.swift
+//  MultichannelWidget
 //
-//  Created by asharijuang on 17/12/19.
+//  Created by Rahardyan Bisma on 16/07/20.
 //
 
 #if os(iOS)
 import UIKit
 #endif
 import Foundation
-import QiscusCoreAPI
+import QiscusCore
 import SwiftyJSON
 import UserNotifications
 
@@ -22,7 +22,7 @@ class QismoManager {
     private var username : String = ""
     private var avatarUrl: String = ""
     var network : QismoNetworkManager!
-    var qiscus : QiscusCoreAPI!
+    var qiscus : QiscusCore!
     var qiscusServer = QiscusServer(url: URL(string: "https://api.qiscus.com")!, realtimeURL: "", realtimePort: 80)
     var deviceToken : String = "" // save device token for 1st time or before login
     
@@ -36,14 +36,17 @@ class QismoManager {
     func clear() {
         self.userID = ""
         self.username = ""
-        self.qiscus.signOut()
+        self.qiscus.clearUser { (error) in
+            print("Qiscus clear user succeeded")
+        }
         SharedPreferences.removeRoomId()
     }
     
     func setup(appID: String, server : QiscusServer? = nil) {
         self.appID = appID
-        self.qiscus = QiscusCoreAPI.init(withAppId: appID, server: qiscusServer)
-        self.network = QismoNetworkManager(QiscusCoreAPI: self.qiscus)
+        self.qiscus = QiscusCore()
+        self.qiscus.setup(AppID: appID)
+        self.network = QismoNetworkManager(qiscusCore: self.qiscus)
         if let _server = server {
             self.qiscusServer = _server
         }
@@ -83,7 +86,7 @@ class QismoManager {
             
             // check device token
             if !self.deviceToken.isEmpty {
-                self.qiscus.register(deviceToken: self.deviceToken, isDevelopment: false, onSuccess: { (success) in
+                self.qiscus.shared.registerDeviceToken(token: self.deviceToken, isDevelopment: false, onSuccess: { (success) in
                     if success { self.deviceToken = "" }
                 }) { (error) in
                     //
@@ -162,3 +165,4 @@ class QismoManager {
     
     
 }
+
